Integrate google calendar API.
Currently fetches all upcoming events form a given account that has given OAuth2.0 access to the app. Im not sure if my account will persist on your guys' systems, so it may ask for your own OAuth2.0 permission. If so, just sign in with one of your google accounts and it'll do the rest.