Add deletion of user when logging out.
This was done by implementing a global user var. This would have been necessary eventually anyways if we wanted the app to actually do a bunch of stuff with things such as the physician's ID. So now we have an access to the logged in or registered user as a global var in constants.swift. Note, however, that this was more of a dirty fix... it may be better to hold this global var in its own separate swift class for handling things more related to the logged in user like maybe log-in time and such (since @Deven you said that we would need to log lots of info or something).